+++ src/Util/TypeUtil.php
@@ -15,9 +15,9 @@ namespace Rekalogika\Mapper\Util;
 
 use DaveLiddament\PhpLanguageExtensions\Friend;
 use DaveLiddament\PhpLanguageExtensions\NamespaceVisibility;
+use Rekalogika\Mapper\Contracts\MixedType;
 use Rekalogika\Mapper\Exception\InvalidArgumentException;
 use Rekalogika\Mapper\Exception\MapperReturnsUnexpectedValueException;
-use Rekalogika\Mapper\Model\MixedType;
 use Rekalogika\Mapper\Tests\UnitTest\Util\TypeUtil2Test;
 use Rekalogika\Mapper\Tests\UnitTest\Util\TypeUtilTest;
 use Rekalogika\Mapper\TypeResolver\TypeResolver;